requirements.txt
opencv-python
numpy
openvino
deep_sort_realtime
PyYAML
